pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = "Payment service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.5.0",
    "colorlog>=6.9.0",
    "django>=5.1.7",
    "django-debug-toolbar>=5.0.1",
    "django-environ>=0.12.0",
    "django-extensions>=3.2.3",
    "django-ninja>=1.4.0",
    "faker>=37.1.0",
    "fief-client[fastapi]>=0.20.0",
    "psycopg[binary]>=3.2.5",
    "requests>=2.32.3",
    "sentry-sdk[django]>=2.22.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
